[US1018][Transactions] - unit tests for visibility of transactions

+++ src/main/java/org/hajecsdb/graphs/transactions/transactionalGraph/TGraph.java
@@ -11,7 +11,7 @@ public interface TGraph {
     Optional<Node> getNodeById(long id);
     Set<Node> getAllNodes();
     Node deleteNode(long id);
-    void deletePropertyFromNode(int nodeId, String propertyKey);
+    void deletePropertyFromNode(long nodeId, String propertyKey);
     Relationship createRelationship(long startNodeId, long endNodeId, Label label);
     Relationship setPropertyToRelationship(long relationshipId, Property property);
     Relationship deleteRelationship(long id);
